pyproject.toml
[project]
name = "sistema-bancario-novo"
version = "0.1.0"
description = ""
authors = [
    {name = "leandrosfreitas",email = "leandrosanfreitas@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.12.3,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "fastapi[standard] (>=0.120.0,<0.121.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "bcrypt (==3.2.2)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
alembic = "^1.17.0"

